requirements.txt
numpy
matplotlib
pandas
ipython
pyserial